+++ src/index.html
@@ -9,16 +9,11 @@
 </head>
 <body>
   <header class="header">
-    <a href="#" class="title">ODOT APP</a>
-    <nav class="nav">
-      <ul>
-        <li><a href="#" class="nav-item">#</a></li>
-      </ul>
-    </nav>
+    <nav class="nav"></nav>
   </header>
   <main>
     <section>
-      <div>
+      <div class="container">
         <ul id="todo-list" class="todo-list"></ul>
       </div>
     </section>
