+++ src/style.css
@@ -5,8 +5,23 @@ body {
   background-color: #e6e6e6;
 }
 
-li {
+.checkbox {
+  vertical-align: middle;
+  font-size: 24px;
+  display: flex;
+  
+}
+
+.todo-item {
   list-style-type: none;
+  color: #000;
+  font-weight: bold;
+  font-size: 24px;
+  box-shadow: 0px 5px 15px rgba(0, 0, 0, 0.25);
+  padding: 10px;
+  margin: 5px 0;
+  border-radius: 8px;
+  width: 100%;
 }
 
 .header {
@@ -29,4 +44,42 @@ li {
 
 .nav-item{
   font-size: 24px;
+}
+
+.container {
+  align-items: center;
+  margin: 20% auto;
+  padding-top: 20px;
+  background-color: #fff;
+  border-radius: 10px;
+  box-shadow: 0px 10px 10px rgba(0, 0, 0, 0.25);
+}
+
+.completed {
+  color: red;
+  font-weight: bold;
+  font-size: 24px;
+}
+
+.todo-list {
+  width: 100%;
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 20px; 
+}
+
+#add-todo {
+  padding: 10px;
+  border: none;
+  border-radius: 5px;
+  font-size: 24px;
+  margin-bottom: 10px;
+  font-weight: bold;
+  box-shadow: 5px 10px 10px rgba(0, 0, 0, 0.25);
+}
+
+.todo-item:hover {
+  background-color: #f0f0f0;
 }
